+++ src/main/java/com/example/constructionsite/machine/repository/MachineRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MachineRepository extends JpaRepository<MachineEntity, Long> {
+public interface MachineRepository extends JpaRepository<MachineEntity, Integer> {
   boolean existsByName(String name);
 }
